CDAP-8032 configurable retry policy for service unavailable errors
Adding a way to configure a retry policy for programs and using
that policy when a service is unavailable. Each program type has
its own default policy that can be set in the cdap configuration.
The default policy can be overridden for a specific program run
using runtime arguments.

ServiceDiscoverer, StreamWriter, Admin, and DatasetContext.getDataset
methods have been changed to use the retry policy.